fix(nn): pin four more fused recurrences against FMA contraction; green the tree
Same defect as the EMA update, found across the fused inference paths after
merging main: GLA, DeltaNet, GatedDeltaNet, RGLRU (Griffin) and HGRN each claim
bit-exactness against their dispatch path and each failed it on this host.

The cause is identical in all of them. The dispatch path builds the state update
out of separate backend ops — a Mul, then an Add — so each product is rounded
before it is added. The fused path writes the same recurrence as one expression,
a*h + b, which the compiler contracts to FMADD on arm64 and generally does not on
amd64. The claim therefore passed CI and failed on Apple silicon. GLA's comment
asserted "no FMA" outright; that is not something a comment can assert, only
something the code can enforce.

Only the state-update product needed rounding. The output dots are deliberately
left alone: those go through a backend matmul on the dispatch side, so the fused
loop has to match whatever the matmul itself does — pinning them broke parity
rather than fixing it, which is why each site was measured one change at a time
instead of rounding everything on principle.

Also documents three exported symbols so apicheck (§V19) goes green. NSABranches
lost its godoc in the previous commit when a helper type was inserted between the
comment and the function; QMatMul and GradientBoostingRegressor.Predict were
undocumented on main.

`go test -short ./...` is now clean across the whole tree — no remaining failures.

+++ nn/deltanet.go
@@ -90,8 +90,13 @@ func DeltaNet(ctx *backend.Context, q, k, v, beta *tensor.Tensor) (*tensor.Tenso
 							p += S[base+c] * krow[c]
 						}
 						d := bt * (vrow[r] - p)
+						// The product is rounded EXPLICITLY. The dispatch path builds this
+						// state update from separate backend ops, each rounding its own
+						// result, while d*krow[c] added in one expression is what the
+						// compiler contracts to FMADD on arm64 — so the parity claim held
+						// on amd64 CI and failed on Apple silicon.
 						for c := range dk {
-							S[base+c] += d * krow[c]
+							S[base+c] += float64(d * krow[c])
 						}
 					}
 				}
